use untitled documents instead of write
use the untitled scheme for the uri instead of writing a file if it doesn't exist. this has the advantage of making VSCode handle all the actual writing of the file, its directory path, etc. It also means that we don't create a file for the user if they didn't want it—they have to confirm its existence by explicity saving.